feat: [GALI-1966] PR comments
