+++ workbench-core/accounts/src/services/accountService.ts
@@ -7,8 +7,7 @@ import { GetItemCommandOutput } from '@aws-sdk/client-dynamodb';
 import { AwsService, resourceTypeToKey, uuidWithLowercasePrefix } from '@aws/workbench-core-base';
 import Boom from '@hapi/boom';
 import _ from 'lodash';
-import Account from '../models/account';
-import { AccountSchema } from '../models/accountSchema';
+import { Account, AccountSchema } from '../models/account';
 
 export default class AccountService {
   private _aws: AwsService;
