+++ workbench-core/accounts/src/services/costCenterService.ts
@@ -12,7 +12,7 @@ import {
   uuidWithLowercasePrefix
 } from '@aws/workbench-core-base';
 import Boom from '@hapi/boom';
-import Account from '../models/account';
+import { Account } from '../models/account';
 import CostCenter from '../models/costCenter';
 import CreateCostCenterRequest from '../models/createCostCenterRequest';
 import AccountService from './accountService';
